+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # NEXT-RELEASE
 - Added `get_default(name)` function to scripting. It allows lookup of constants like the `Shopware/src/Core/Defaults.php`
+- Fixed import of "To-Many-Associations" when the value is null it will be ignored instead of added to the entity
 
 # v0.4.0
 - "To-One-Association" values are now imported correctly